% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.parttree}
\alias{plot.parttree}
\title{Plot decision tree partitions}
\usage{
\method{plot}{parttree}(
  object,
  raw = TRUE,
  border = "black",
  fill_alpha = 0.3,
  expand = TRUE,
  jitter = FALSE,
  xlab = NULL,
  ylab = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{object}{A \link{parttree} data frame.}

\item{raw}{Logical. Should the raw (original) data points be plotted too?
Default is TRUE.}

\item{border}{Colour of the partition borders (edges). Default is "black". To
remove the borders altogether, specify as \code{NA}.}

\item{fill_alpha}{Numeric in the range \verb{[0,1]}. Alpha transparency of the
filled partition rectangles. Default is \code{0.3}.}

\item{expand}{Logical. Should the partition limits be expanded to to meet the
edge of the plot axes? Default is \code{TRUE}. If \code{FALSE}, then the partition
limits will extend only until the range of the raw data.}

\item{jitter}{Logical. Should the raw points be jittered? Default is \code{FALSE}.
Only evaluated if \code{raw = TRUE}.}

\item{xlab, ylab}{Character string(s). Custom labels for the x and y axes.}

\item{add}{Logical. Add to an existing plot? Default is \code{FALSE}.}

\item{...}{Additional arguments passed down to
\code{\link[graphics]{tinyplot}}\link{tinyplot}.}
}
\value{
No return value, called for side effect of producing a plot.
}
\description{
Provides a plot method for parttree objects.
}
\examples{
## rpart tree example
library("rpart")
rp = rpart(Kyphosis ~ Start + Age, data = kyphosis)
pt = parttree(rp)

## simple plot
plot(pt)

## removing the (recursive) partition borders helps to emphasise overall fit
plot(pt, border = NA)

## customize further by passing extra options to (tiny)plot
plot(
    pt,
    border  = NA,                                     # no partition borders
    pch     = 19,                                     # filled points
    alpha   = 0.6,                                    # point transparency
    grid    = TRUE,                                   # background grid
    palette = "classic",                              # new colour palette
    xlab    = "Topmost vertebra operated on",         # custom x title
    ylab    = "Patient age (months)",                 # custom y title
    main    = "Tree predictions: Kyphosis recurrence" # custom title
)
}
