% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parttree.R
\name{parttree}
\alias{parttree}
\title{Convert a decision tree to a data table of partition coordinates}
\usage{
parttree(tree)
}
\arguments{
\item{tree}{An rpart object, or rpart-compatible object (e.g. using parsnip)}
}
\value{
A data table comprising seven columns: the leaf node, its path, a set of coordinates understandable to ggplot2 (i.e. xmin, xmax, ymin, ymax), and a final column corresponding to the predicted value for that leaf.
}
\description{
Extracts the terminal leaf nodes of a decision tree with one or two predictor variables. These leaf nodes are then converted into a data table, such that there partition
}
\details{
This function can be used with a regression or classification tree containing one or (at most) two continuous predictors.
}
\examples{
library(rpart)
parttree(rpart(Species ~ Sepal.Width + Petal.Width, data=iris))
}
\keyword{partition}
