% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parttree.R
\name{parttree}
\alias{parttree}
\title{Convert a decision tree into a data frame of partition coordinates}
\usage{
parttree(tree, keep_as_dt = FALSE)
}
\arguments{
\item{tree}{An \code{\link[rpart]{rpart.object}}, or an object of compatible
type (e.g. a decision tree constructed via the `parsnip` or `mlr3`
front-ends).}

\item{keep_as_dt}{Logical. The function relies on `data.table` for internal
data manipulation. But it will coerce the final return object into a
regular data frame (default behaviour) unless the user specifies `TRUE`.}
}
\value{
A data frame comprising seven columns: the leaf node, its path, a set
  of coordinates understandable to `ggplot2` (i.e. xmin, xmax, ymin, ymax),
  and a final column corresponding to the predicted value for that leaf.
}
\description{
Extracts the terminal leaf nodes of a decision tree with one or
  two predictor variables. These leaf nodes are then converted into a data
  frame, where each row represents a partition that can easily be plotted in
  coordinate space.
}
\details{
This function can be used with a regression or classification tree
  containing one or (at most) two continuous predictors.
}
\examples{
library(rpart)
parttree(rpart(Species ~ Petal.Length + Petal.Width, data=iris))
}
\seealso{
\code{\link{geom_parttree()}}, \code{\link[rpart]{rpart}}.
}
